Refactored Code and added Cofig Files
