+++ ESP32_Client_V1.0/src/Distributor.cpp
@@ -12,8 +12,6 @@
  */
 
 #include "Distributor.h"
-#include "Device.h"
-
 
 Distributor::Distributor(InstrumentController* ptrInstrumentController)
 {
