+++ ESP32_Client_V1.0/src/Instruments/InstrumentController.h
@@ -1,8 +1,6 @@
 #pragma once
 
 #include "Constants.h"
-#include "Device.h"
-#include "Platform.h"
 #include <cstdint>
 using std::int8_t;
